The Traverser API has loops() to be an int, but in SimpleTraverser, we are using a short to save memory. That is 37000+ loops able to be represented and thus, 99.999999% of use cases.

+++ gremlin-core/src/main/java/com/tinkerpop/gremlin/process/Traverser.java
@@ -63,7 +63,7 @@ public interface Traverser<T> extends Serializable, Comparable<Traverser<T>> {
      *
      * @return The number of times the traverser has gone through a loop
      */
-    public short loops();
+    public int loops();
 
     /**
      * A traverser may represent a grouping of traversers to allow for more efficient data propagation.
